Move `isFeatureFlagActive` function into the same file for other feature flag methods
